requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.0
camelot-py==0.11.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
distro==1.9.0
et-xmlfile==1.1.0
filelock==3.15.4
ghostscript==0.7
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.8
JPype1==1.5.0
numpy==2.1.0
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pdfminer.six==20240706
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyparsing==3.1.2
pypdf==4.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==73.0.1
six==1.16.0
soupsieve==2.6
tabulate==0.9.0
tk==0.1.0
tqdm==4.66.5
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
